requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.1=pyhd8ed1ab_0
aom=3.6.1=h59595ed_0
blas=1.0=mkl
brotlicffi=1.1.0.0=py312h2ec8cdc_2
bzip2=1.0.8=h5eee18b_6
c-ares=1.34.5=hb9d3cd8_0
ca-certificates=2025.7.14=hbd8a1cb_0
certifi=2025.7.14=pyhd8ed1ab_0
cffi=1.17.1=py312h1fdaa30_1
charset-normalizer=3.4.2=pyhd8ed1ab_0
cpython=3.12.11=py312hd8ed1ab_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-opencl=12.9.19=0
cuda-runtime=11.8.0=0
cuda-version=12.9=3
expat=2.7.1=h6a678d5_0
ffmpeg=4.4.2=gpl_hdf48244_113
filelock=3.18.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.15.0=h7e30c49_1
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
freetype=2.13.3=ha770c72_1
giflib=5.2.2=h5eee18b_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py312h5eee18b_0
gnutls=3.7.9=hb077bed_0
grpcio=1.73.1=py312hf7a0c6d_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.7.0=pyhe01879c_1
intel-openmp=2022.0.1=h06a4308_3633
jinja2=3.1.6=py312h06a4308_0
joblib=1.5.1=pyhd8ed1ab_0
lame=3.100=h7b6447c_0
lcms2=2.17=h717163a_0
ld_impl_linux-64=2.44=h1423503_1
lerc=4.0.0=h6a678d5_0
libabseil=20250512.1=cxx17_hba17884_0
libasprintf=0.25.1=h8e693c7_0
libblas=3.9.0=16_linux64_mkl
libcblas=3.9.0=16_linux64_mkl
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.14.1.1=4
libcurand=10.3.10.19=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.24=h86f0d12_0
libdrm=2.4.125=hb9d3cd8_0
libegl=1.7.0=h5eee18b_2
libexpat=2.7.1=hecca717_0
libffi=3.4.6=h2dba641_1
libfreetype=2.13.3=ha770c72_1
libfreetype6=2.13.3=h48d6fc4_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgettextpo=0.25.1=h5888daf_0
libgfortran=15.1.0=h69a702a_3
libgfortran5=15.1.0=hcea5267_3
libgl=1.7.0=h5eee18b_2
libglvnd=1.7.0=h5eee18b_2
libglx=1.7.0=h5eee18b_2
libgomp=15.1.0=h767d61c_3
libgrpc=1.73.1=h1e535eb_0
libiconv=1.18=h4ce23a2_1
libidn2=2.3.8=ha4ef2c3_0
libjpeg-turbo=3.1.0=hb9d3cd8_0
liblapack=3.9.0=16_linux64_mkl
liblzma=5.8.1=hb9d3cd8_2
libnpp=11.8.0.86=0
libnsl=2.0.1=hb9d3cd8_1
libnvjitlink=12.1.105=0
libnvjpeg=11.9.0.86=0
libpciaccess=0.18=h5eee18b_0
libpng=1.6.50=h943b412_0
libprotobuf=6.31.1=h9ef548d_1
libre2-11=2025.06.26=h7064273_1
libsqlite=3.50.2=hee844dc_2
libstdcxx=15.1.0=h8f9b012_3
libstdcxx-ng=15.1.0=h4852527_3
libtasn1=4.20.0=hb9d3cd8_0
libtiff=4.7.0=hf01ce69_5
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h4f16b4b_2
libvpx=1.13.1=h6a678d5_0
libwebp=1.6.0=h9635ea4_0
libwebp-base=1.6.0=hd42ef1d_0
libxcb=1.17.0=h9b100fa_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.8=h4bc477f_0
libzlib=1.3.1=hb9d3cd8_2
llvm-openmp=15.0.7=h0cdce71_0
markdown=3.8.2=pyhd8ed1ab_0
markupsafe=3.0.2=py312h5eee18b_0
mkl=2022.1.0=hc2b9512_224
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py312h06a4308_0
ncurses=6.5=h2d0b736_3
nettle=3.9.1=h7ab15ed_0
networkx=3.5=pyhe01879c_0
numpy=2.3.1=py312h33ff503_1
ocl-icd=2.3.3=hb9d3cd8_0
opencl-headers=2025.06.13=h5888daf_0
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.3=h5fbd93e_0
openssl=3.5.1=h7b32b05_0
p11-kit=0.24.1=hc5aa10d_0
packaging=25.0=pyh29332c3_1
pillow=11.3.0=py312h80c1187_0
pip=25.1.1=pyh8b19718_0
protobuf=6.31.1=py312h7e1e8fc_0
pthread-stubs=0.4=hb9d3cd8_1002
pycparser=2.22=pyh29332c3_1
pysocks=1.7.1=py312h06a4308_0
python=3.12.11=h9e4cc4f_0_cpython
python_abi=3.12=6_cp312
pytorch=2.5.1=py3.12_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py312h5eee18b_0
re2=2025.06.26=h9925aae_1
readline=8.2=h5eee18b_0
requests=2.32.4=py312h06a4308_0
scikit-learn=1.7.0=py312h7a48858_1
scipy=1.16.0=py312hf734454_0
setuptools=80.9.0=pyhff2d567_0
sh=2.2.2=pyh707e725_1
svt-av1=1.4.1=hcb278e6_0
sympy=1.14.0=pyh2585a3b_105
tensorboard=2.20.0=pyhe01879c_0
tensorboard-data-server=0.7.0=py312hda17c39_2
threadpoolctl=3.6.0=pyhecae5ae_0
tk=8.6.13=noxft_hd72426e_102
torchaudio=2.5.1=py312_cu118
torchtriton=3.1.0=py312
torchvision=0.20.1=py312_cu118
tqdm=4.67.1=py312he106c6f_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=py312h06a4308_0
wayland=1.24.0=hdac8c69_0
wayland-protocols=1.45=hd8ed1ab_0
werkzeug=3.1.3=pyhd8ed1ab_1
wheel=0.45.1=py312h06a4308_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-libxext=1.3.6=h9b100fa_0
xorg-libxfixes=6.0.1=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zipp=3.23.0=pyhd8ed1ab_0
zstd=1.5.7=hb8e6e7a_2
